requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
python-dateutil>=2.8.0
pydantic>=2.0.0
